[RISCV] Implement hardening guards for jumps, returns and tail returns
This patch adds support for `unimp` guard instructions following return,
tail, and unconditional jump instructions, as a hardening measure. The
guard instructions are emitted during the MC-layer encoding stage for
object-code emission and during instruction printing when generating
textual assembly code output.

The computation of instruction sizes has been updated to account for the
size of the guards that will be emitted during the late-stage expansion.