[Clang][RISCV] Add support for double_check attribute
The ``clang::double_check`` attribute is used to annotate `if` statements
where the condition expression should be checked twice, as a mitigation for
hardware attacks. This attribute is currently RISC-V specific and only
applies to `if` statements with a binary expression where the operator is
one of ``==``, ``!=``, ``>``, ``>=``, ``<``, or ``<=`` and the operands are
of integer type.